requirements.txt
selenium
pandas
pymongo
PyYAML
pydantic
playwright